+++ scripts/release.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 
-set -e # exit when error
+# set -e # exit when error
+
+ROOT_DIR=$(pwd)
 
 if ! npm owner ls | grep -q "$(npm whoami)"
 then
-  printf "Release: Not an owner of the npm repo, ask for it\n"
+  printf "Release: Not an owner of the npm repo, ask for it.\n"
   exit 1
 fi
 
@@ -24,12 +26,55 @@ if [[ $# -eq 0 ]] ; then
   exit 1
 fi
 
+# Bump project version
 mversion $1
+
+NEW_VERSION=$(node -p -e "require('$ROOT_DIR/package.json').version")
+CHECK_NEW_TAG_REMOTE=$(git tag | grep "v$NEW_VERSION")
+
+if [ "$CHECK_NEW_TAG_REMOTE" != "" ]; then
+    printf "The tag $NEW_VERSION is already released.\n"
+    exit 1
+fi
+
+REMOTE_TAGS=$(git tag --list 'v*[0-9.]')
+LATEST_TAG=$(printf "%s\n" "v$NEW_VERSION" "${REMOTE_TAGS}" | sort --version-sort | tail -1)
+
+if [ "$LATEST_TAG" != "v$NEW_VERSION" ]; then
+    printf "You are trying to release an older version $NEW_VERSION than the one on remote $LATEST_TAG.\n"
+    exit 1
+fi
+
+# Update README ToC
 doctoc README.md
-git commit -am "$(json -f package.json version)"
-git tag v`json -f package.json version`
+
+# Update and tag Git project
+git commit -am "${NEW_VERSION}"
+git tag -a v${NEW_VERSION}
 git push
 git push --tags
 npm publish
 
-printf "Release: please update the CHANGELOG manually\n"
+
+# Docker
+printf "Release: build Docker image, tag it (latest and ${NEW_VERSION}) and push it to Docker hub\n"
+
+cd docker
+
+# Build image
+docker build -t vvoyer/selenium-standalone . --rm
+
+# Login to default Docker hub (https://hub.docker.com/)
+docker login
+
+# Tag image and push it to Docker hub
+for tag in ${NEW_VERSION} latest
+do
+    docker tag vvoyer/selenium-standalone vvoyer/selenium-standalone:${tag}
+    docker push vvoyer/selenium-standalone:${tag}
+done
+
+cd ..
+
+# Reminder to manually update project history file
+printf "Release: please update the HISTORY manually\n"